refactor reporter names, prepend legacy report with v3
