+++ ExtentReports/Views/Html/Device.cshtml
