+++ ExtentReports/Views/Html/Partials/Log.cshtml
