--- ExtentReports/Views/Html2/IHtml2Marker.cs
@@ -1,6 +1,0 @@
-﻿namespace AventStack.ExtentReports.Views.Html2
-{
-    internal interface IHtml2Marker
-    {
-    }
-}
