+++ ExtentReports/Views/V3Html/IV3HtmlMarker.cs
@@ -1,0 +1,6 @@
+﻿namespace AventStack.ExtentReports.Views.V3Html
+{
+    internal interface IV3HtmlMarker
+    {
+    }
+}
