+++ ExtentReports/Views/V3Html/Test/V3Test.cshtml
@@ -94,7 +94,7 @@
 			<!-- search -->
 		</div>
 	</section>
-	@Include("Charts")
+	@Include("V3Charts")
 	<div class='subview-left left'>
 		<div class='view-summary'>
 			<ul id='test-collection' class='test-collection'>
